requirements.txt
python-telegram-bot>=20.0
Pillow>=10.0
rlottie-python>=1.0.0
python-dotenv>=1.0.0
imageio-ffmpeg>=0.4.0
discord.py>=2.0
